requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
prometheus-fastapi-instrumentator
pip install joblib